+++ app/src/main/java/com/study/thesuperiorstanislav/decisiontheorylabs/lab1/Lab1Fragment.kt
@@ -29,7 +29,7 @@ import android.util.Log
 import android.view.Window
 import com.jjoe64.graphview.series.PointsGraphSeries
 import com.study.thesuperiorstanislav.decisiontheorylabs.utils.File.FileReader
-import com.study.thesuperiorstanislav.decisiontheorylabs.utils.GraphHelper
+import com.study.thesuperiorstanislav.decisiontheorylabs.utils.Graph.GraphHelper
 import kotlinx.android.synthetic.main.app_bar_main.*
 import org.jetbrains.anko.onComplete
 import org.jetbrains.anko.uiThread
